pyproject.toml
[project]
name = "silo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.97.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
]
